Was probably not a good idea
They can probably implement it better for their own use-case or it can 
be added later.
